try both v4 and v6?
