+++ tests/test_util/has_multicast.hpp
@@ -20,21 +20,14 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "has_ipv4_multicast.hpp"
-
-#include <algorithm>
-
-#include "util/network/get_interfaces.hpp"
-#include "util/platform.hpp"
+#ifndef TEST_UTIL_HAS_IPV4_MULTICAST_HPP
+#define TEST_UTIL_HAS_IPV4_MULTICAST_HPP
 
 namespace test_util {
 
-bool has_ipv4_multicast() {
-    // See if any interface has multicast ipv4
-    auto ifaces = NUClear::util::network::get_interfaces();
-    return std::any_of(ifaces.begin(), ifaces.end(), [](const auto& iface) {
-        return iface.ip.sock.sa_family == AF_INET && iface.flags.multicast;
-    });
-}
+bool has_ipv4_multicast();
+bool has_ipv6_multicast();
 
 }  // namespace test_util
+
+#endif  // TEST_UTIL_HAS_IPV4_MULTICAST_HPP
